socket/requirements.txt
python-socketio===5.5.2
eventlet===0.33.0
SQLAlchemy===1.4.32
configparser===5.2.0
sqlalchemy-migrate===0.13.0
psycopg2===2.9.3
typing_extensions